pasando los primeros tests

+++ src/HTTPResponsev2.h
@@ -87,7 +87,7 @@ static const char * CONTENT_STRING = "CONTENT-";
 struct http_response {
     char http_version; 
     char headers[MAX_HEADERS_RESPONSE_LENGTH];
-    uint16_t header_content_length;
+    uint32_t header_content_length;
     char content_types[MAX_TYPES][MAX_TYPES_LEN]; 
     char code_reason[MAX_REASON_LENGTH];
     int status_code;
@@ -117,7 +117,7 @@ struct http_res_parser {
   uint16_t i_encoding;
   uint16_t i_type; 
   uint16_t i_c_encoding; 
-  int content_length;
+  uint32_t content_length;
   uint16_t transfer_encodings; 
   uint16_t content_encodings; 
   uint16_t content_types;
